+++ CHANGELOG.md
@@ -3,3 +3,4 @@ Papyros Shell CHANGELOG
 
 * [Dashboard] Fix some issues with the Music widget
 * Fix the shell not closing when running as a Wayland compositor
+* [Desktop] Improve the hot corners